+++ src/engine/animations.rs
@@ -17,7 +17,7 @@ pub trait TimingFunction {
 impl TimingFunction for Easing {
     fn value_at(&self, t: f32) -> f32 {
         let Easing { x1, x2, y1, y2 } = *self;
-        bezier_easing_function(x1, x2, y1, y2, t)
+        bezier_easing_function(x1 as f64, x2 as f64, y1 as f64, y2 as f64, t as f64) as f32
     }
 }
